frontend/requirements.txt
streamlit
pandas
numpy
streamlit-cookies-manager
python-dotenv
requests
openai
os
